scripts/oracle-instance-launcher/requirements.txt
oci>=2.150.0
python-dotenv>=1.0.1
